Remove unnecessary use of Option for wrap/filter.
These shouldn't be represented via Option as their values are read
unconditionally anyway, so we can rely on their defaults if they are not
explicitly specified.

Anisotropic max is only used for anisotropic
filtering methods, but its value is also read unconditionally in that
case so we can just use its default as well.